LayoutTests/platform/blackberry/editing/pasteboard/paste-blockquote-before-blockquote-expected.txt

CONSOLE MESSAGE: line 20: FAILED: pasting singleNestedNodeToCopy DID keep the newline within the blockquote
This tests pasting a singly nested blockquote before, and outside of, another blockquote. This test does not work outside of DRT.

line 1
line 2
line 3
line 4
line 5
